+++ packages/nitro-protocol/src/contract/transaction-creators/nitro-adjudicator.ts
@@ -1,10 +1,10 @@
 // @ts-ignore
 import {TransactionRequest} from 'ethers/providers';
-import {Interface} from 'ethers/utils';
+import {Interface, Signature} from 'ethers/utils';
 import NitroAdjudicatorArtifact from '../../../build/contracts/NitroAdjudicator.json';
 import {getChannelId} from '../channel';
 import {encodeOutcome, Outcome} from '../outcome';
-import {hashState, State} from '../state';
+import {getFixedPart, hashAppPart, hashState, State} from '../state';
 
 // TODO: Currently we are setting some arbitrary gas limit
 // to avoid issues with Ganache sendTransaction and parsing BN.js
@@ -35,4 +35,53 @@ export function createPushOutcomeTransaction(
   ]);
 
   return {data, gasLimit: GAS_LIMIT};
+}
+
+export function concludePushOutcomeAndTransferAllArgs(
+  states: State[],
+  signatures: Signature[],
+  whoSignedWhat: number[]
+): any[] {
+  // Sanity checks on expected lengths
+  if (states.length === 0) {
+    throw new Error('No states provided');
+  }
+  const {participants} = states[0].channel;
+  if (participants.length !== signatures.length) {
+    throw new Error(
+      `Participants (length:${participants.length}) and signatures (length:${signatures.length}) need to be the same length`
+    );
+  }
+
+  const lastState = states.reduce((s1, s2) => (s1.turnNum >= s2.turnNum ? s1 : s2), states[0]);
+  const largestTurnNum = lastState.turnNum;
+  const fixedPart = getFixedPart(lastState);
+  const appPartHash = hashAppPart(lastState);
+
+  const outcomeBytes = encodeOutcome(lastState.outcome);
+
+  const numStates = states.length;
+
+  return [
+    largestTurnNum,
+    fixedPart,
+    appPartHash,
+    outcomeBytes,
+    numStates,
+    whoSignedWhat,
+    signatures,
+  ];
+}
+
+export function createConcludePushOutcomeAndTransferAllTransaction(
+  states: State[],
+  signatures: Signature[],
+  whoSignedWhat: number[]
+): TransactionRequest {
+  return {
+    data: NitroAdjudicatorContractInterface.functions.concludePushOutcomeAndTransferAll.encode(
+      concludePushOutcomeAndTransferAllArgs(states, signatures, whoSignedWhat)
+    ),
+    gasLimit: GAS_LIMIT,
+  };
 }
